+++ tm-asset-api/src/main/groovy/cn/edu/bnuz/bell/asset/RoomCommand.groovy
@@ -53,6 +53,11 @@ class RoomCommand {
      */
     String note
 
+    /**
+     * 标签
+     */
+    List<Long> labels
+
     String department
     String placeType
 }
