+++ tm-asset-api/src/main/groovy/cn/edu/bnuz/bell/asset/RoomLifeCommand.groovy
@@ -57,4 +57,9 @@ class RoomLifeCommand {
      * 备注
      */
     String note
+
+    /**
+     *分拆的场地信息
+     */
+    List<RoomCommand> rooms
 }
